+++ src/main/java/com/excellenthealthSolution/pharmacy/asset/prescriber/dao/ConsultationDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConsultationDao extends JpaRepository<Consultations, Integer> {
+public interface ConsultationDao extends JpaRepository<Consultations, Long> {
 }
