+++ src/main/java/com/excellenthealthSolution/pharmacy/asset/suppliers/dao/SupplierContactPersonDao.java
@@ -3,5 +3,5 @@ package com.excellenthealthSolution.pharmacy.asset.suppliers.dao;
 import com.excellenthealthSolution.pharmacy.asset.suppliers.entity.SupplierContactPerson;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SupplierContactPersonDao extends JpaRepository<SupplierContactPerson, Integer> {
+public interface SupplierContactPersonDao extends JpaRepository<SupplierContactPerson, Long> {
 }
