+++ src/main/java/com/excellenthealthSolution/pharmacy/security/controller/RoleController.java
@@ -1,7 +1,7 @@
 package com.excellenthealthSolution.pharmacy.security.controller;
 
-import com.excellenthealthSolution.pharmacy.security.entity.Role;
-import com.excellenthealthSolution.pharmacy.security.service.RoleService;
+import com.brothers.premasiri.security.entity.Role;
+import com.brothers.premasiri.security.service.RoleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -32,13 +32,13 @@ public class RoleController {
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public String roleView(@PathVariable("id") Integer id, Model model) {
+    public String roleView(@PathVariable("id") Long id, Model model) {
         model.addAttribute("roleDetail", roleService.findById(id));
         return "role/role-detail";
     }
 
     @RequestMapping(value = "/edit/{id}", method = RequestMethod.GET)
-    public String editRoleFrom(@PathVariable("id") Integer id, Model model) {
+    public String editRoleFrom(@PathVariable("id") Long id, Model model) {
         model.addAttribute("role", roleService.findById(id));
         model.addAttribute("addStatus", false);
         return "role/addRole";
@@ -58,7 +58,7 @@ public class RoleController {
     public String addRole(@Valid @ModelAttribute Role role, BindingResult result, Model model) {
         System.out.println(role);
         if (result.hasErrors()) {
-            for (FieldError error : result.getFieldErrors()) {
+            for ( FieldError error : result.getFieldErrors()) {
                 System.out.println(error.getField() + ": " + error.getDefaultMessage());
             }
             model.addAttribute("addStatus", true);
@@ -70,7 +70,7 @@ public class RoleController {
     }
 
     @RequestMapping(value = "/remove/{id}", method = RequestMethod.GET)
-    public String removeRole(@PathVariable Integer id) {
+    public String removeRole(@PathVariable Long id) {
         roleService.delete(id);
         return "redirect:/role";
     }
@@ -81,3 +81,4 @@ public class RoleController {
         return "role/role-detail";
     }
 }
+
