+++ src/main/java/com/excellenthealthSolution/pharmacy/security/dao/RoleDao.java
@@ -1,5 +1,7 @@
 package com.excellenthealthSolution.pharmacy.security.dao;
 
+
+
 import com.excellenthealthSolution.pharmacy.security.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -7,5 +9,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface RoleDao extends JpaRepository<Role, Integer> {
+public interface RoleDao extends JpaRepository< Role, Long> {
 }
